MAJOR: Add new scenario test framework with many file-based operations, Rewrite RenameFileSystemAction to handle all cases, DeleteFileSystemAction still buggy (does not catch all cases); MINOR: add rename-tests that failed in real-life

+++ src/org/syncany/operations/actions/DeleteFileSystemAction.java
@@ -24,12 +24,17 @@ public class DeleteFileSystemAction extends FileSystemAction {
 		// Normal case: locally known file to be deleted
 		else {
 			if (!fileAsExpected(fileVersion1)) {
-				createConflictFile(fileVersion2);				
+				if (fileExists(fileVersion1)) {
+					createConflictFile(fileVersion2);	
+				}
 			}
 		}
 			
 		File toDeleteFileOnDisk = getAbsolutePathFile(fileVersion2.getFullName());
-		toDeleteFileOnDisk.delete();
+		
+		if (toDeleteFileOnDisk.exists()) {
+			toDeleteFileOnDisk.delete();
+		}
 	}
 
 	@Override
